# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.Result:
    properties:
      data:
        type: object
      error:
        $ref: '#/definitions/api.ResultError'
        type: object
      status:
        type: string
    type: object
  api.ResultError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  model.Pagination:
    properties:
      current:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  model.ResultSelect:
    properties:
      label:
        type: string
      value:
        type: object
    type: object
...
  model.TableListData:
    properties:
      list:
        type: object
      pagination:
        $ref: '#/definitions/model.Pagination'
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: 862860000@qq.com
    name: API Support
    url: https://github.com/nilorg/naas
  description: NilOrg认证授权服务Api详情.
  license:
    name: GNU General Public License v3.0
    url: https://github.com/nilorg/naas/blob/master/LICENSE
  termsOfService: https://github.com/nilorg
  title: NilOrg认证授权服务
  version: "1.0"
paths:
  /casbin/resource/{resource_server_id}/actions:
    get:
      consumes:
      - application/json
      parameters:
      - description: resource server id
        in: path
        name: resource_server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 获取资源服务器动作翻页列表
      tags:
      - Casbin
  /casbin/resource/{resource_server_id}/menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: resource server id
        in: path
        name: resource_server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 获取资源服务器菜单翻页列表
      tags:
      - Casbin
  /casbin/resource/{resource_server_id}/routes:
    get:
      consumes:
      - application/json
      parameters:
      - description: resource server id
        in: path
        name: resource_server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 获取资源服务器路由翻页列表
      tags:
      - Casbin
  /casbin/role/{role_code}/resource/{resource_server_id}/actions:
    get:
      consumes:
      - application/json
      parameters:
      - description: role code
        in: path
        name: role_code
        required: true
        type: string
      - description: resource server id
        in: path
        name: resource_server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 根据角色获取资源服务器动作
      tags:
      - Casbin
  /casbin/role/{role_code}/resource/{resource_server_id}/menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: role code
        in: path
        name: role_code
        required: true
        type: string
      - description: resource server id
        in: path
        name: resource_server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 根据角色获取资源服务器菜单翻页列表
      tags:
      - Casbin
  /casbin/role/{role_code}/resource/{resource_server_id}/routes:
    get:
      consumes:
      - application/json
      parameters:
      - description: role code
        in: path
        name: role_code
        required: true
        type: string
      - description: resource server id
        in: path
        name: resource_server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 根据角色获取资源服务器路由翻页列表
      tags:
      - Casbin
  /casbin/role/{role_code}/resource_actions:
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色Code
        in: path
        name: role_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 添加资源动作
      tags:
      - Casbin
  /casbin/role/{role_code}/resource_menus:
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色Code
        in: path
        name: role_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 添加资源菜单
      tags:
      - Casbin
  /casbin/role/{role_code}/resource_routes:
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色Code
        in: path
        name: role_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 添加资源路由
      tags:
      - Casbin
  /common/select:
    get:
      consumes:
      - application/json
      description: |-
        organization:组织
        resource_server:资源服务器
        role:角色
        oauth2_scope:OAuth2Scope
      parameters:
      - description: 查询参数
        enum:
        - organization
        - resource_server
        - role
        - oauth2_scope
        in: query
        name: q
        required: true
        type: string
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.ResultSelect'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 查询角色
      tags:
      - Common（通用）
  /common/tree:
    get:
      consumes:
      - application/json
      description: resource_menu:菜单
      parameters:
      - description: 查询参数
        enum:
        - resource_menu
        in: query
        name: q
        required: true
        type: string
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.ResultSelect'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 查询角色
      tags:
      - Common（通用）
...
securityDefinitions:
  OAuth2AccessCode:
    authorizationUrl: http://localhost:8080/oauth2/authorize
    flow: accessCode
    scopes:
      email: ' 用户emial'
      openid: ' 用户openid'
      phone: ' 用户手机号'
      profile: ' 用户资料'
    tokenUrl: http://localhost:8080/oauth2/token
    type: oauth2
swagger: "2.0"